Prevent `Pty::pts()` from attaching via `O_NOCTTY`
If a process doesn't have a controlling terminal (for example if it is
in a new session) then `open(ptsname())` will automatically make the
slave the controlling terminal of the process.

Later on, `command.spawn(&pts)`s call to `ioctl_tiocsctty()` within
`session_leader()` within `command.pre_exec()` will fail as with EPERM
since the parent process (and session) is alread using it as its
controlling terminal.

Use `O_NOCTTY` to prevent this.